kenter_energy/requirements.txt
requests==2.31.0
python-dateutil==2.8.2
paho-mqtt==1.6.1 